Remove files of the target kernel under /boot
After the rollback, kernel / bootloader files of target kernel have
been kept in the /boot parition. As easier solution of this,
I put cleaning of this file automatically when the --clean option is
used. As we require run of r-u-t with the --clean option after the
rollback, it is for now safe solution.

To ensure that we remove only files that are created during the
upgrade and will not remove kernel files that it is available
under dual boot (unexpected, but just to be sure), it is checked
version of target kernel and in case `kernelpath` of this version
doesn't exist already, information is stored under `rollback_dir`
in the `.target-kernel` file. Only in case the data is stored in the
file, kernel files of target kernel will be cleaned.

+++ redhat_upgrade_tool/rollback/bootloader.py
@@ -4,6 +4,8 @@ import shutil
 import platform
 from subprocess import CalledProcessError, call
 
+from redhat_upgrade_tool.rollback import rollback_dir
+
 try:
     from redhat_upgrade_tool import grub_conf_file
     from redhat_upgrade_tool.util import check_call
@@ -93,6 +95,20 @@ def restore_boot(release=platform.release()):
 def clean_snapshot_boot_files():
     for fmt in _SNAP_BOOT_FILES:
         path = os.path.join("/boot", fmt.format("snapshot"))
+        if os.path.isfile(path):
+            os.remove(path)
+
+
+def clean_target_boot_files():
+    # get target_kernel
+    target_kernel_file = os.path.join(rollback_dir,".target-kernel")
+    if not os.path.isfile(target_kernel_file):
+        # nothing to do
+        return
+    with open(target_kernel_file ,"rb") as kerf:
+        target_kernel = kerf.read()
+    for fmt in _SNAP_BOOT_FILES:
+        path = os.path.join("/boot", fmt.format(target_kernel))
         if os.path.isfile(path):
             os.remove(path)
